frontends: move frontends/android/goserver go backend/mobileserver
An iOS app in frontends/ios would use the same package, just with a
different build target.

frontend/android: move ANDROID_SDK_ROOT/ANDROID_NDK_HOME definitions

Otherwise one can't run `make` inside `mobileserver/`.
