+++ android-env.mk.inc
@@ -1,0 +1,2 @@
+export ANDROID_SDK_ROOT := /opt/android-sdk
+export ANDROID_NDK_HOME := /opt/android-sdk/ndk/21.2.6472646
