+++ backend/mobileserver/.gitignore
@@ -1,0 +1,2 @@
+/mobileserver-sources.jar
+/mobileserver.aar
