+++ backend/mobileserver/README.md
@@ -1,0 +1,4 @@
+# BitBox Android
+
+`make build-android` builds android.aar, which can be imported as a module into an Android project
+(Android Studio).
