+++ frontends/android/BitBoxApp/.gitignore
@@ -9,11 +9,11 @@
 /.idea/assetWizardSettings.xml
 .DS_Store
 /build
-/goserver/build
+/mobileserver/build
 /captures
 .externalNativeBuild
 
 # Autogenerated web assets.
 /app/src/main/assets/web
-# Autogenerated goserver output
-/goserver/goserver.aar
+# Autogenerated mobileserver output
+/mobileserver/mobileserver.aar
