+++ frontends/android/BitBoxApp/app/build.gradle
@@ -35,5 +35,5 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation "android.arch.lifecycle:extensions:1.1.1"
     implementation "android.arch.lifecycle:viewmodel:1.1.1"
-    implementation project(path: ':goserver')
+    implementation project(path: ':mobileserver')
 }
