--- frontends/android/BitBoxApp/goserver/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('goserver.aar'))
