+++ frontends/android/BitBoxApp/mobileserver/build.gradle
@@ -1,0 +1,2 @@
+configurations.maybeCreate("default")
+artifacts.add("default", file('mobileserver.aar'))
