+++ frontends/android/BitBoxApp/settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':goserver'
+include ':app', ':mobileserver'
