--- frontends/android/goserver/.gitignore
@@ -1,2 +1,0 @@
-/goserver-sources.jar
-/goserver.aar
