--- frontends/android/goserver/README.md
@@ -1,4 +1,0 @@
-# BitBox Android Library
-
-`make build` builds android.aar, which can be imported as a module into an Android project
-(Android Studio).
